chore: identify causal consistency prose tests
